Simplifications and cleanup.
