+++ x/project/internal/types/msgs.go
@@ -8,9 +8,8 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/auth"
 	"github.com/ixofoundation/ixo-blockchain/x/did"
 	"github.com/ixofoundation/ixo-blockchain/x/did/exported"
-	"github.com/spf13/viper"
-
 	"github.com/ixofoundation/ixo-blockchain/x/ixo"
+	"github.com/spf13/viper"
 )
 
 const (
@@ -45,6 +44,17 @@ type MsgCreateProject struct {
 	Data       json.RawMessage `json:"data" yaml:"data"`
 }
 
+func NewMsgCreateProject(senderDid did.Did, projectData json.RawMessage,
+	projectDid did.Did, pubKey string) MsgCreateProject {
+	return MsgCreateProject{
+		TxHash:     "",
+		SenderDid:  senderDid,
+		ProjectDid: projectDid,
+		PubKey:     pubKey,
+		Data:       projectData,
+	}
+}
+
 func (msg MsgCreateProject) ToStdSignMsg(fee int64) auth.StdSignMsg {
 	chainID := viper.GetString(flags.FlagChainID)
 	accNum, accSeq := uint64(0), uint64(0)
@@ -68,9 +78,9 @@ func (msg MsgCreateProject) Route() string { return RouterKey }
 
 func (msg MsgCreateProject) ValidateBasic() sdk.Error {
 	// Check that not empty
-	if valid, err := CheckNotEmpty(msg.PubKey, "PubKey"); !valid {
+	if valid, err := checkNotEmpty(msg.PubKey, "PubKey"); !valid {
 		return err
-	} else if valid, err := CheckNotEmpty(msg.ProjectDid, "ProjectDid"); !valid {
+	} else if valid, err := checkNotEmpty(msg.ProjectDid, "ProjectDid"); !valid {
 		return err
 	}
 
@@ -86,12 +96,7 @@ func (msg MsgCreateProject) ValidateBasic() sdk.Error {
 	if !found {
 		return sdk.ErrInternal("missing evaluatorPayPerClaim in project doc")
 	}
-	var evaluatorPayPerClaimStr string
-	err = json.Unmarshal(evaluatorPayPerClaimBz, &evaluatorPayPerClaimStr)
-	if err != nil {
-		return sdk.ErrInternal(err.Error())
-	}
-	_, err = sdk.ParseCoins(evaluatorPayPerClaimStr)
+	_, err = sdk.ParseCoins(withoutQuotes(string(evaluatorPayPerClaimBz)))
 	if err != nil {
 		return sdk.ErrInternal("evaluatorPayPerClaim should be valid coins")
 	}
@@ -101,12 +106,7 @@ func (msg MsgCreateProject) ValidateBasic() sdk.Error {
 	if !found {
 		return sdk.ErrInternal("missing claimerPayPerClaim in project doc")
 	}
-	var claimerPayPerClaimStr string
-	err = json.Unmarshal(claimerPayPerClaimBz, &claimerPayPerClaimStr)
-	if err != nil {
-		return sdk.ErrInternal(err.Error())
-	}
-	_, err = sdk.ParseCoins(claimerPayPerClaimStr)
+	_, err = sdk.ParseCoins(withoutQuotes(string(claimerPayPerClaimBz)))
 	if err != nil {
 		return sdk.ErrInternal("claimerPayPerClaim should be valid coins")
 	}
@@ -158,14 +158,23 @@ type MsgUpdateProjectStatus struct {
 	Data       UpdateProjectStatusDoc `json:"data" yaml:"data"`
 }
 
+func NewMsgUpdateProjectStatus(senderDid did.Did, updateProjectStatusDoc UpdateProjectStatusDoc, projectDid did.Did) MsgUpdateProjectStatus {
+	return MsgUpdateProjectStatus{
+		TxHash:     "",
+		SenderDid:  senderDid,
+		ProjectDid: projectDid,
+		Data:       updateProjectStatusDoc,
+	}
+}
+
 func (msg MsgUpdateProjectStatus) Type() string  { return TypeMsgUpdateProjectStatus }
 func (msg MsgUpdateProjectStatus) Route() string { return RouterKey }
 
 func (msg MsgUpdateProjectStatus) ValidateBasic() sdk.Error {
 	// Check that not empty
-	if valid, err := CheckNotEmpty(msg.ProjectDid, "ProjectDid"); !valid {
+	if valid, err := checkNotEmpty(msg.ProjectDid, "ProjectDid"); !valid {
 		return err
-	} else if valid, err := CheckNotEmpty(msg.SenderDid, "SenderDid"); !valid {
+	} else if valid, err := checkNotEmpty(msg.SenderDid, "SenderDid"); !valid {
 		return err
 	}
 
@@ -199,13 +208,22 @@ type MsgCreateAgent struct {
 	Data       CreateAgentDoc `json:"data" yaml:"data"`
 }
 
+func NewMsgCreateAgent(txHash string, senderDid did.Did, createAgentDoc CreateAgentDoc, projectDid did.Did) MsgCreateAgent {
+	return MsgCreateAgent{
+		ProjectDid: projectDid,
+		TxHash:     txHash,
+		SenderDid:  senderDid,
+		Data:       createAgentDoc,
+	}
+}
+
 func (msg MsgCreateAgent) Type() string  { return TypeMsgCreateAgent }
 func (msg MsgCreateAgent) Route() string { return RouterKey }
 func (msg MsgCreateAgent) ValidateBasic() sdk.Error {
 	// Check that not empty
-	if valid, err := CheckNotEmpty(msg.ProjectDid, "ProjectDid"); !valid {
+	if valid, err := checkNotEmpty(msg.ProjectDid, "ProjectDid"); !valid {
 		return err
-	} else if valid, err := CheckNotEmpty(msg.SenderDid, "SenderDid"); !valid {
+	} else if valid, err := checkNotEmpty(msg.SenderDid, "SenderDid"); !valid {
 		return err
 	}
 
@@ -247,13 +265,22 @@ type MsgUpdateAgent struct {
 	Data       UpdateAgentDoc `json:"data" yaml:"data"`
 }
 
+func NewMsgUpdateAgent(txHash string, senderDid did.Did, updateAgentDoc UpdateAgentDoc, projectDid did.Did) MsgUpdateAgent {
+	return MsgUpdateAgent{
+		ProjectDid: projectDid,
+		TxHash:     txHash,
+		SenderDid:  senderDid,
+		Data:       updateAgentDoc,
+	}
+}
+
 func (msg MsgUpdateAgent) Type() string  { return TypeMsgUpdateAgent }
 func (msg MsgUpdateAgent) Route() string { return RouterKey }
 func (msg MsgUpdateAgent) ValidateBasic() sdk.Error {
 	// Check that not empty
-	if valid, err := CheckNotEmpty(msg.ProjectDid, "ProjectDid"); !valid {
+	if valid, err := checkNotEmpty(msg.ProjectDid, "ProjectDid"); !valid {
 		return err
-	} else if valid, err := CheckNotEmpty(msg.SenderDid, "SenderDid"); !valid {
+	} else if valid, err := checkNotEmpty(msg.SenderDid, "SenderDid"); !valid {
 		return err
 	}
 
@@ -296,14 +323,23 @@ type MsgCreateClaim struct {
 	Data       CreateClaimDoc `json:"data" yaml:"data"`
 }
 
+func NewMsgCreateClaim(txHash string, senderDid did.Did, createClaimDoc CreateClaimDoc, projectDid did.Did) MsgCreateClaim {
+	return MsgCreateClaim{
+		ProjectDid: projectDid,
+		TxHash:     txHash,
+		SenderDid:  senderDid,
+		Data:       createClaimDoc,
+	}
+}
+
 func (msg MsgCreateClaim) Type() string  { return TypeMsgCreateClaim }
 func (msg MsgCreateClaim) Route() string { return RouterKey }
 
 func (msg MsgCreateClaim) ValidateBasic() sdk.Error {
 	// Check that not empty
-	if valid, err := CheckNotEmpty(msg.ProjectDid, "ProjectDid"); !valid {
+	if valid, err := checkNotEmpty(msg.ProjectDid, "ProjectDid"); !valid {
 		return err
-	} else if valid, err := CheckNotEmpty(msg.SenderDid, "SenderDid"); !valid {
+	} else if valid, err := checkNotEmpty(msg.SenderDid, "SenderDid"); !valid {
 		return err
 	}
 
@@ -344,14 +380,23 @@ type MsgCreateEvaluation struct {
 	Data       CreateEvaluationDoc `json:"data" yaml:"data"`
 }
 
+func NewMsgCreateEvaluation(txHash string, senderDid did.Did, createEvaluationDoc CreateEvaluationDoc, projectDid did.Did) MsgCreateEvaluation {
+	return MsgCreateEvaluation{
+		ProjectDid: projectDid,
+		TxHash:     txHash,
+		SenderDid:  senderDid,
+		Data:       createEvaluationDoc,
+	}
+}
+
 func (msg MsgCreateEvaluation) Type() string  { return TypeMsgCreateEvaluation }
 func (msg MsgCreateEvaluation) Route() string { return RouterKey }
 
 func (msg MsgCreateEvaluation) ValidateBasic() sdk.Error {
 	// Check that not empty
-	if valid, err := CheckNotEmpty(msg.ProjectDid, "ProjectDid"); !valid {
+	if valid, err := checkNotEmpty(msg.ProjectDid, "ProjectDid"); !valid {
 		return err
-	} else if valid, err := CheckNotEmpty(msg.SenderDid, "SenderDid"); !valid {
+	} else if valid, err := checkNotEmpty(msg.SenderDid, "SenderDid"); !valid {
 		return err
 	}
 
@@ -390,16 +435,23 @@ type MsgWithdrawFunds struct {
 	Data      WithdrawFundsDoc `json:"data" yaml:"data"`
 }
 
+func NewMsgWithdrawFunds(senderDid did.Did, data WithdrawFundsDoc) MsgWithdrawFunds {
+	return MsgWithdrawFunds{
+		SenderDid: senderDid,
+		Data:      data,
+	}
+}
+
 func (msg MsgWithdrawFunds) Type() string  { return TypeMsgWithdrawFunds }
 func (msg MsgWithdrawFunds) Route() string { return RouterKey }
 
 func (msg MsgWithdrawFunds) ValidateBasic() sdk.Error {
 	// Check that not empty
-	if valid, err := CheckNotEmpty(msg.SenderDid, "SenderDid"); !valid {
+	if valid, err := checkNotEmpty(msg.SenderDid, "SenderDid"); !valid {
 		return err
-	} else if valid, err := CheckNotEmpty(msg.Data.ProjectDid, "ProjectDid"); !valid {
+	} else if valid, err := checkNotEmpty(msg.Data.ProjectDid, "ProjectDid"); !valid {
 		return err
-	} else if valid, err := CheckNotEmpty(msg.Data.RecipientDid, "RecipientDid"); !valid {
+	} else if valid, err := checkNotEmpty(msg.Data.RecipientDid, "RecipientDid"); !valid {
 		return err
 	}
